Change files to TS
