+++ package.json
@@ -39,6 +39,7 @@
     "@types/node-sass": "~4.11.1",
     "@types/react": "~16.9.47",
     "@types/react-dom": "~16.9.8",
+    "@types/react-router-dom": "~5.1.8",
     "@typescript-eslint/eslint-plugin": "~4.18.0",
     "@typescript-eslint/parser": "~4.18.0",
     "acorn": "~7.1.1",
